+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changelog
 
 ### Fixes
 
+- **[IR]** Fix exclusions not inherited from parent graphs
 - **[IR/Native]** Fix parameter type equality check in native compilations in constructor-injected class factories.
 - **[IR/Compat]** Fix compatibiltiy for source-less diagnostics in Kotlin `2.3.20-Beta1`.